start of testing

+++ VirtualRouter/Background/Packet/src/Egress.cpp
@@ -21,8 +21,8 @@ Egress::~Egress()
 bool Egress::sendPacket(const ByteString& base256Str) 
 {
   std::vector<unsigned char> packet_data = base256ToBytes(base256Str);
-  int packet_length = packet_data.size();
-  if (pcap_sendpacket(pcap_handle, packet_data.data(), packet_length) != 0) 
+  size_t packet_length = packet_data.size();
+  if (pcap_sendpacket(pcap_handle, packet_data.data(), static_cast<int>(packet_length)) != 0) 
   {
       std::cerr << "Error sending packet: " << pcap_geterr(pcap_handle) << std::endl;
       return false;
