+++ VirtualRouter/Background/Protocols/Global/src/Ospf.cpp
@@ -5,4 +5,4 @@ Ospf::Ospf() {}
 }
 
 Protocol::Ospf* currentOspf;
-std::map<int, std::shared_ptr<Protocol::Ospf>> ospfList;
+std::map<uint16_t, std::shared_ptr<Protocol::Ospf>> ospfList;
